ref(issues): Share navigation count styles
