+++ static/app/views/navigation/secondary/sections/issues/issueViews/issueViewQueryCount.tsx
@@ -1,17 +1,11 @@
-import styled from '@emotion/styled';
-
-import {Tag} from '@sentry/scraps/badge';
-import {Text} from '@sentry/scraps/text';
-
 import type {PageFilters} from 'sentry/types/core';
 import {getUtcDateString} from 'sentry/utils/dates';
 import {useLocation} from 'sentry/utils/useLocation';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {createIssueViewFromUrl} from 'sentry/views/issueList/issueViews/createIssueViewFromUrl';
 import {useFetchIssueCounts} from 'sentry/views/issueList/queries/useFetchIssueCounts';
+import {IssueCount} from 'sentry/views/navigation/secondary/sections/issues/issueCount';
 import type {IssueView} from 'sentry/views/navigation/secondary/sections/issues/issueViews/issueViews';
-
-const TAB_MAX_COUNT = 99;
 
 const constructCountTimeFrame = (
   timeFilters: PageFilters['datetime']
@@ -71,33 +65,5 @@ export function IssueViewQueryCount({view, isActive}: IssueViewQueryCountProps) 
     return null;
   }
 
-  return (
-    <StyledTag variant="muted" data-issue-view-query-count>
-      <Text variant="muted" size="xs" align="center" tabular>
-        {count > TAB_MAX_COUNT ? `${TAB_MAX_COUNT}+` : count}
-      </Text>
-    </StyledTag>
-  );
+  return <IssueCount count={count} />;
 }
-const StyledTag = styled(Tag)`
-  border: 1px solid ${p => p.theme.tokens.border.neutral.muted};
-  background-color: ${p => p.theme.tokens.background.primary};
-  padding: 0 ${p => p.theme.space.xs};
-  justify-content: end;
-
-  opacity: 0;
-  transform: scale(0.95);
-
-  @keyframes fadeIn {
-    from {
-      opacity: 0;
-      transform: scale(0.95);
-    }
-    to {
-      opacity: 1;
-      transform: scale(1);
-    }
-  }
-
-  animation: fadeIn 0.1s ease-in-out forwards;
-`;
