+++ static/app/views/navigation/secondary/sections/issues/issuesSecondaryNavigation.tsx
@@ -10,13 +10,11 @@ import {Tooltip} from '@sentry/scraps/tooltip';
 import {t, tct} from 'sentry/locale';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {makeAutomationBasePathname} from 'sentry/views/automations/pathnames';
-import {
-  INBOX_COUNT_MAX,
-  useInboxIssueCount,
-} from 'sentry/views/issueList/queries/useInboxIssueCount';
+import {useInboxIssueCount} from 'sentry/views/issueList/queries/useInboxIssueCount';
 import {ISSUE_TAXONOMY_CONFIG} from 'sentry/views/issueList/taxonomies';
 import {usePrimaryNavigation} from 'sentry/views/navigation/primaryNavigationContext';
 import {SecondaryNavigation} from 'sentry/views/navigation/secondary/components';
+import {IssueCount} from 'sentry/views/navigation/secondary/sections/issues/issueCount';
 import {IssueViews} from 'sentry/views/navigation/secondary/sections/issues/issueViews/issueViews';
 
 function InboxCountBadge() {
@@ -26,11 +24,7 @@ function InboxCountBadge() {
     return null;
   }
 
-  return (
-    <Badge variant="muted">
-      {count > INBOX_COUNT_MAX ? `${INBOX_COUNT_MAX}+` : count}
-    </Badge>
-  );
+  return <IssueCount count={count} />;
 }
 
 export function IssuesSecondaryNavigation() {
